[feat] 세션 Disconnection 정보를 로직서버 및 클라이언트 모두에게 Broadcast하도록 변경; 패킷 헤더 수정

+++ IocpLib/Fundamentals/Enumerations.h
@@ -11,7 +11,7 @@ enum HostTypeEnum : uint16
 	MIDDLEMAN_SERVER,
 };
 
-enum ActionTypeEnum : uint32
+enum ActionTypeEnum : uint8
 {
 	UNDEFINED,
 	MOVE,
